minor issue

+++ transport_classes/Ground_transport.h
@@ -9,8 +9,8 @@ protected:
       third_time = 0;
   double second_time = 0;
   std::string type = "Наземный транспорт";
-  double rest_count = range / move_until_rest;
-  double final_time = (range / speed) + time_calc();
+  double rest_count = (double )range / move_until_rest;
+  double final_time = ((double )range / speed) + time_calc();
 };
 
 //    final_time = (range/speed) + all_rest;
